\name{synthetic_data}
\alias{synthetic_data}
\title{Synthetic data set to test the functions of \emph{RolWinMulCor}. 
}
\description{
The data set \code{\link{synthetic_data}} contains five columns, the first one is the time and the next four columns are sinusoidal time series that contains two periodical signals (at 11 and 51, with a phase of zero and amplitude of 1 --please note that these quantities are dimensionless) 'contaminated' by Gaussian noise (with mean of 0 and standard deviation of 0.25) for the intervals 1-100 (signal at 11) and 200-400 (signal at 51) and Gaussian noise (with mean of 0 and standard deviation of 1) otherwise. 
}
\usage{data(synthetic_data)
}
\format{One file in ASCII format and columns are separated by spaces. 
}
\source{
Author's own production (Josué M. Polanco-Martínez). 
}
\keyword{dataset1}
